Modify storage example to showcase high-memory remote calls

+++ lib/tinykvm/common.hpp
@@ -60,8 +60,8 @@ namespace tinykvm
 		uint32_t max_cow_mem = 0;
 		uint32_t stack_size = 1600UL << 10; /* 1600KB */
 		uint32_t reset_free_work_mem = 0; /* reset_to() */
-		uint32_t dylink_address_hint = 0x200000; /* 2MB */
-		uint32_t heap_address_hint = 0;
+		uint64_t dylink_address_hint = 0x200000; /* 2MB */
+		uint64_t heap_address_hint = 0;
 		uint64_t vmem_base_address = 0;
 		std::string_view binary = {};
 		std::vector<VirtualRemapping> remappings {};
